requirements.txt
# ── Airflow ───────────────────────────────────────────────────────────────────
apache-airflow==2.9.1

# ── Web scraping (extract task) ───────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── Document extraction ───────────────────────────────────────────────────────
pdfplumber

# ── Data processing ───────────────────────────────────────────────────────────
pandas

# ── Vector store, embeddings & LLM orchestration ─────────────────────────────
langchain
langchain-anthropic
langchain-community
langchain-text-splitters
langgraph
sentence-transformers
faiss-cpu

# ── Database / config ─────────────────────────────────────────────────────────
pymongo
pyyaml
python-dotenv

# ── Dashboard (Streamlit service) ─────────────────────────────────────────────
streamlit
plotly
folium
streamlit-folium
altair
